Bugfix with 'null' index, and support for spaces in list keys
	* (M) lib/yangtypes.py -- fix an issue whereby a key that was 'null'
				would result in _set() not being called correctly, particularly
				generally the test was 'if val' - which mean that where the
				value was explicitly set to (e.g., 0) this broke. Particularly
				this was a problem when testing against default values - since
				0 == False.
	* (M) lib/yangtypes.py -- as per issue #28 (thanks Mark Paine).
				Add support for using kwargs to add items to lists where the
				user might expect spaces in the list index AND there is a
				compound key. The add() function can be used with keyword
				arguments, and a new _item() function can be used to retrieve
				using keyword arguments. The old approach is left in place for
				backwards compatibility. (NB: deserialising code may need
				to use this new approach where spaces are expected too)
				This commit adds an enhancement to add() such that it returns
				the value that was added such that code can be compressed.

+++ tests/run.sh
@@ -3,6 +3,8 @@
 # Find where we are meant to run the tests
 TESTDIR="$(cd -P "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 
+DELENV=false
+
 # Build a virtual environment to do the tests in
 cd $TESTDIR/..
 rm -rf $TESTDIR/pyvirtualenv $TESTDIR/../dist $TESTDIR/../build $TESTDIR/../pyangbind.egg-info
@@ -36,4 +38,6 @@ $TESTDIR/yang_tests.sh
 $TESTDIR/xpath/xpath_tests.sh
 $TESTDIR/serialise/serialise_tests.sh
 
-rm -rf $TESTDIR/pyvirtualenv $TESTDIR/../dist $TESTDIR/../build $TESTDIR/../pyangbind.egg-info
+if [ "$DELENV" == "true" ]; then
+    rm -rf $TESTDIR/pyvirtualenv $TESTDIR/../dist $TESTDIR/../build $TESTDIR/../pyangbind.egg-info
+fi
